requirements.txt
streamlit
pymupdf
python-docx
Pillow